ETL/.history/requirements_20250709202011.txt
pandas>=1.0.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
pyarrow
tqdm
A-Lib